rework: make custom themes easier
