+++ oxytail-theme-defaults/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "oxytail-theme-dark"
-description = "Dark theme for oxytail-base"
+name = "oxytail-theme-defaults"
+description = "A set of uncolored sane defaults for creating simple themes."
 license = "MIT"
 version = "0.1.0"
 edition = "2021"
@@ -11,4 +11,3 @@ rust-version = "1.75"
 [dependencies]
 floem = { git = "https://github.com/lapce/floem", rev = "e304d30f8771a28788904d64464d8fd192d07439" }
 oxytail-base = { path = "../oxytail-base" }
-oxytail-theme-defaults = { path = "../oxytail-theme-defaults" }
